Create Basic MapView in MainActivity
This commit adds the permisisons for Google Maps to the AndroidManifest
file and adds a MapView in activity_main.xml.
It also changes MainActivity.java to allow the MapView.

For this MapView to work a file must be created: res/values/apikeys.xml
In this file the string release_maps_key and debug_maps_key should be
saved.