Resolve two NullPointerException Crashes
A user who installed through the Google Play Store encountered an issue
where the line:
-   GeoPoint oldGeo = new GeoPoint((int)
(locationManager.getLastKnownLocation(LocationManager.NETWORK_PROVIDER).getLatitude()
* 1E6), (int)
(locationManager.getLastKnownLocation(LocationManager.NETWORK_PROVIDER).getLongitude()
* 1E6));
Crashed Popup Places, presumably because their was no Last Known
Location for the Network Provider (the device was new).
I've put this code in a try/catch and implemented a request for a single
location update if it fails.

On every device disabling the background service could cause a crash
because of this line:
mNotificationManager.cancelAll();
This could fail because mNotificationManager had never been initialized.
I solved this by only executing this code if mNotification is not null.